connect the feature module navigation graph to the app module navigation graph
1. connect the feature module navigation graph to the app module navigation graph
2. remove app module fragments
3. add bottom navigation view selector
