+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:theme="@style/Theme.TaskerAndroid"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
