--- app/src/main/java/com/example/tasker_android/ui/mypage/MyPageFragment.kt
@@ -1,17 +1,0 @@
-package com.example.tasker_android.ui.mypage
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import com.example.tasker_android.base.BaseFragment
-import com.example.tasker_android.databinding.FragmentMyPageBinding
-
-class MyPageFragment : BaseFragment<FragmentMyPageBinding>() {
-    override fun getFragmentBinding(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-    ): FragmentMyPageBinding {
-        return FragmentMyPageBinding.inflate(inflater, container, false)
-    }
-
-
-}
