+++ app/src/main/res/drawable/bottom_nav_selector.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:state_checked="true"
+        android:color="@color/black" />
+    <item
+        android:state_checked="false"
+        android:color="@color/gray_d9" />
+</selector>
