+++ feature/collection/src/main/res/layout/fragment_collection.xml
@@ -9,6 +9,6 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:text="collection fragment" />
 
 </FrameLayout>
